Introducing a new OVSDB table to keep track of OVS-DPDK hw-accel features.
Updating the OVSDB schema with a new table to keep track of hardware
acceleration features that are relevant to OVS. Each hardware device will
populate a row in this table with their properties.

The new hw table is introduced due to following reasons

1) Hardware acceleration in userspace datapath is handled completely via
DPDK. It is not necessary that control interface/tools are present in kernel
for querying or configuring underlying hardware. Since DPDK offers only
APIs to access these details, OVSDB is used to display it to the
management layer/orchestrator.

2) Similar to DPDK parameters, hardware acceleration does have parameters such
as number of virtual ports to use in OVS, allocated memory, hw features to
enable and etc. OVSDB stores those information for orchestrator similar to
how its been done for DPDK parameters.

Change-Id: I567961808861be9246c0b02bd7c3dbeddb5a7ce7
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>