Helper function to get original inport from mbuf.
OVS-DPDK installs flows in hardware using in-port. The in-port must be
an accelerated type, to trigger the flow installation in dpif layer.
Sometimes OVS-DPDK need to do the flow install on non-accelerated
virtual ports such as 'tunnel'. The helper function is used for those
cases to find original hardware in-port.

Change-Id: I6919d71d4b9f8c08806127b012d26fd6d49d26be
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>