Backport of standard hardware acceleration netdev functions
Backporting the standard netdev flow offload functions that are defined
as part of kernel hardware acceleration implementation. Most of the
function are same as kernel implementation, except the 'flow_put'. OVS-DPDK
may need to install flow in hardware even when the inport is  non-
accelerated port such as tunnel.
The 'flow-put' is modified to find and install a flow on original accelerated
port even when inport is not a accelerated type.

Change-Id: I72443999f09a8d17f84ec7a0fc9340854623b5b8
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>