Helper functions for ovs flow + actions sanity checking.
Adding helper functions in OVS-DPDK hardware acceleration framework to
perform basic validation of datapath flow + actions before trying to program
the hardware. The scope of these functions are very limited to reduce its
overhead on flow installation logic. Currently these functions check if the
datapath protocol is being supported in hardware or not.
Also it verifies if the actions can be handled completely in the same device.

Change-Id: Ic5ab9e722c0b95f129409cd5962957f9ca9ea674
Signed-off-by: Chandran, Sugesh <sugesh.chandran@intel.com>